requirements.txt
ttkbootstrap>=1.10.1
pillow>=10.3.0
tkinterdnd2>=0.4.2